requirements.txt
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
pydantic==1.10.13
